Move types to higher level
