+++ src/components/EditListResultViewer/EditListResultViewer.tsx
@@ -5,7 +5,7 @@ import { t } from '../../services';
 import { getVisibleColumnsKey } from '../../utils';
 
 import { ConfigureQuery } from '../ConfigureQuery';
-import { STATUS_VALUES, VISIBILITY_VALUES } from '../../pages/editlist/types';
+import { STATUS_VALUES, VISIBILITY_VALUES } from '../../interfaces';
 import css from '../../pages/createlist/components/MainCreateListForm/MainCreateListForm.module.css';
 
 type EditListResultViewerProps = {
